+++ camverything/src/main/java/com/multi/camp/login/mypageController.java
@@ -30,7 +30,17 @@ public class mypageController {
 		this.res_service = res_service;
 		this.pro_service = pro_service;
 	}
-
+	
+	@RequestMapping(value = "/main/mypage_update")
+	public String mypage_update() {
+		return "main/mypage_update";
+	}
+	
+	@RequestMapping(value = "/main/mypage")
+	public String mypage() {
+		return "main/mypage_main";
+	}
+	
 	@RequestMapping("/main/mypagereservation")
 	public ModelAndView mypagereservation(HttpSession session) {
 		ModelAndView mav = new ModelAndView("main/mypage_reservation");
